Move out HAF and TFA build to GitHub workflow and consume the bins locally (#100)
## Description

For SBSA, move out HAF and TFA build to GitHub workflow and consume the
bins locally.
Eliminates need for local developer to rely on HAF/TFA build steps via a
new `HAF_TFA_BUILD` build flag set to `FALSE` by default.

The GitHub Workflow on Release will run `stuart_build` with
`HAF_TFA_BUILD=TRUE` and publishes the Hafnium and TFA binaries for
consumption as a part of the GitHub release. It will also publish a file
`fip_blob_manifest.json` which is generated using the output of the
fiptool from TFA against fip.bin. This is needed because the fiptool
reports offsets that are necessary to patch up the fip.bin if we want to
employ this method of building SBSA with HAF/TFA bins being pulled down
from an extdep.

The GitHub workflow runs the end-to-end build with `HAF_TFA_BUILD=TRUE`
only on GitHub releases. We will not publish the binaries after building
on arbitrary PR runs, but will still run stuart_build with
`HAF_TFA_BUILD=TRUE`.
On github releases, the version number and sha for the extdep must also
be updated for every release, as the contents of the binaries and the
json manifest, `fip_blob_manifest.json` needs to get updated also.

Once this is merged, a developer can build SBSA like they would
normally, however the default behavior for building the Hafnium and TFA
binaries will be to use the extdep through `stuart_update`. Along with
this, the Post-Build step will now patch the extdep binaries with the
contents of the secure partitions that the local developer has built as
a part of `stuart_build`.
For more advanced use cases, where the developer would need to modify
the secure partition DTS files, or if the developer changes the .fd
files to be larger than the size reported in the
`fip_blob_manifest.json`, then the developer would have to use
`HAF_TFA_BUILD=TRUE` in their `stuart_build` step as an argument.

After this PR is merged, we need to make a formal GitHub release so that
the necessary binaries and artifacts can be published as a part of that
release.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Tested on my fork of patina-qemu by making a release and pulling the
ext_dep to build and run SBSA.

## Integration Instructions

Update Platforms/QemuSbsaPkg/Binaries/haf_tfa_binaries_ext_dep.yaml
every time there is a new release.